Improved validation of digital signature of binaries.
Improved validation of digital signature of binaries. If a valid embedded signature isn't present, we now check for a catalog signature (instead of our previous check: the binary having a pdb present in the microsoft symbol server), this is more secure (as we didn't validate before that the binary pdb we were downloading from the symbol matched with the binary on disk 1:1). I have now removed the 'http' and 'pdb' interaction source and header files, as they are no longer needed.