Added hash double linked list + all kernel driver .text integrity checks
Added hash linked list + all kernel driver .text integrity checks. The linked list (bidirectional) is used to store hashes of different drivers and their identifier. The anticheat driver will also subsequently be integrity checked as it will be in that linked list. If a module is found in PsLoadedModulesList which isn't in the list, it will be appended to the end.